converge virtual-table and module handler
