--- examples/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "/Users/noyman/miniconda3/envs/py3.7/bin/python"
-}
