--- examples/Virtual_table/static_examples/cityio_api_v1.json
@@ -1,1 +1,0 @@
-{"meta":{"id":"3d555f337737ac71f538b1e874a27ca113d52f14200b5d4ee1d67694ed0b4ac1","timestamp":1535379691556,"apiv":"2.1.0"},"header":{"name":"","spatial":{"nrows":0,"ncols":0,"physical_longitude":0,"physical_latitude":0,"longitude":0,"latitude":0,"cellSize":0,"rotation":0},"owner":{"name":"","title":"","institute":""},"block":null,"mapping":null},"grid":[{"rot":180,"type":-2,"x":13,"y":0},{"rot":180,"type":2,"x":12,"y":0},{"rot":180,"type":0,"x":11,"y":0},{"rot":180,"type":-1,"x":10,"y":0},{"rot":180,"type":3,"x":9,"y":0},{"rot":180,"type":0,"x":8,"y":0},{"rot":180,"type":3,"x":7,"y":0},{"rot":180,"type":4,"x":6,"y":0},{"rot":180,"type":5,"x":5,"y":0},{"rot":180,"type":0,"x":4,"y":0},{"rot":180,"type":0,"x":3,"y":0},{"rot":180,"type":4,"x":2,"y":0},{"rot":180,"type":0,"x":1,"y":0},{"rot":180,"type":5,"x":0,"y":0},{"rot":180,"type":-2,"x":13,"y":1},{"rot":180,"type":4,"x":12,"y":1},{"rot":180,"type":-1,"x":11,"y":1},{"rot":180,"type":2,"x":10,"y":1},{"rot":180,"type":-1,"x":9,"y":1},{"rot":180,"type":-1,"x":8,"y":1},{"rot":180,"type":2,"x":7,"y":1},{"rot":180,"type":2,"x":6,"y":1},{"rot":180,"type":5,"x":5,"y":1},{"rot":180,"type":4,"x":4,"y":1},{"rot":180,"type":5,"x":3,"y":1},{"rot":180,"type":5,"x":2,"y":1},{"rot":180,"type":6,"x":1,"y":1},{"rot":180,"type":0,"x":0,"y":1},{"rot":180,"type":-2,"x":13,"y":2},{"rot":180,"type":4,"x":12,"y":2},{"rot":180,"type":-1,"x":11,"y":2},{"rot":180,"type":-1,"x":10,"y":2},{"rot":180,"type":0,"x":9,"y":2},{"rot":180,"type":0,"x":8,"y":2},{"rot":180,"type":5,"x":7,"y":2},{"rot":180,"type":-2,"x":6,"y":2},{"rot":180,"type":-2,"x":5,"y":2},{"rot":180,"type":-2,"x":4,"y":2},{"rot":180,"type":-2,"x":3,"y":2},{"rot":180,"type":-2,"x":2,"y":2},{"rot":180,"type":-2,"x":1,"y":2},{"rot":180,"type":-2,"x":0,"y":2},{"rot":180,"type":-2,"x":13,"y":3},{"rot":180,"type":2,"x":12,"y":3},{"rot":180,"type":0,"x":11,"y":3},{"rot":180,"type":0,"x":10,"y":3},{"rot":180,"type":0,"x":9,"y":3},{"rot":180,"type":1,"x":8,"y":3},{"rot":180,"type":1,"x":7,"y":3},{"rot":180,"type":-2,"x":6,"y":3},{"rot":180,"type":-2,"x":5,"y":3},{"rot":180,"type":-2,"x":4,"y":3},{"rot":180,"type":-2,"x":3,"y":3},{"rot":180,"type":-2,"x":2,"y":3},{"rot":180,"type":-2,"x":1,"y":3},{"rot":180,"type":-2,"x":0,"y":3},{"rot":180,"type":-2,"x":13,"y":4},{"rot":180,"type":1,"x":12,"y":4},{"rot":180,"type":3,"x":11,"y":4},{"rot":180,"type":4,"x":10,"y":4},{"rot":180,"type":4,"x":9,"y":4},{"rot":180,"type":1,"x":8,"y":4},{"rot":180,"type":1,"x":7,"y":4},{"rot":180,"type":-2,"x":6,"y":4},{"rot":180,"type":-2,"x":5,"y":4},{"rot":180,"type":-2,"x":4,"y":4},{"rot":180,"type":-2,"x":3,"y":4},{"rot":180,"type":-2,"x":2,"y":4},{"rot":180,"type":-2,"x":1,"y":4},{"rot":180,"type":-2,"x":0,"y":4},{"rot":180,"type":-2,"x":13,"y":5},{"rot":180,"type":1,"x":12,"y":5},{"rot":180,"type":0,"x":11,"y":5},{"rot":180,"type":5,"x":10,"y":5},{"rot":180,"type":5,"x":9,"y":5},{"rot":180,"type":0,"x":8,"y":5},{"rot":180,"type":1,"x":7,"y":5},{"rot":180,"type":-2,"x":6,"y":5},{"rot":180,"type":-2,"x":5,"y":5},{"rot":180,"type":-2,"x":4,"y":5},{"rot":180,"type":-2,"x":3,"y":5},{"rot":180,"type":-2,"x":2,"y":5},{"rot":180,"type":-2,"x":1,"y":5},{"rot":180,"type":-2,"x":0,"y":5},{"rot":180,"type":-2,"x":13,"y":6},{"rot":180,"type":4,"x":12,"y":6},{"rot":180,"type":0,"x":11,"y":6},{"rot":180,"type":0,"x":10,"y":6},{"rot":180,"type":-1,"x":9,"y":6},{"rot":180,"type":5,"x":8,"y":6},{"rot":180,"type":3,"x":7,"y":6},{"rot":180,"type":-2,"x":6,"y":6},{"rot":180,"type":-2,"x":5,"y":6},{"rot":180,"type":-2,"x":4,"y":6},{"rot":180,"type":-2,"x":3,"y":6},{"rot":180,"type":-2,"x":2,"y":6},{"rot":180,"type":-2,"x":1,"y":6},{"rot":180,"type":-2,"x":0,"y":6},{"rot":180,"type":-1,"x":13,"y":7},{"rot":180,"type":0,"x":12,"y":7},{"rot":180,"type":0,"x":11,"y":7},{"rot":180,"type":4,"x":10,"y":7},{"rot":180,"type":2,"x":9,"y":7},{"rot":180,"type":4,"x":8,"y":7},{"rot":180,"type":-1,"x":7,"y":7},{"rot":180,"type":-2,"x":6,"y":7},{"rot":180,"type":-2,"x":5,"y":7},{"rot":180,"type":-2,"x":4,"y":7},{"rot":180,"type":-2,"x":3,"y":7},{"rot":180,"type":-2,"x":2,"y":7},{"rot":180,"type":-2,"x":1,"y":7},{"rot":180,"type":-2,"x":0,"y":7},{"rot":180,"type":4,"x":13,"y":8},{"rot":180,"type":5,"x":12,"y":8},{"rot":180,"type":0,"x":11,"y":8},{"rot":180,"type":5,"x":10,"y":8},{"rot":180,"type":0,"x":9,"y":8},{"rot":180,"type":0,"x":8,"y":8},{"rot":180,"type":0,"x":7,"y":8},{"rot":180,"type":2,"x":6,"y":8},{"rot":180,"type":2,"x":5,"y":8},{"rot":180,"type":-2,"x":4,"y":8},{"rot":180,"type":-2,"x":3,"y":8},{"rot":180,"type":-2,"x":2,"y":8},{"rot":180,"type":-2,"x":1,"y":8},{"rot":180,"type":-2,"x":0,"y":8},{"rot":180,"type":5,"x":13,"y":9},{"rot":180,"type":0,"x":12,"y":9},{"rot":180,"type":0,"x":11,"y":9},{"rot":180,"type":0,"x":10,"y":9},{"rot":180,"type":0,"x":9,"y":9},{"rot":180,"type":2,"x":8,"y":9},{"rot":180,"type":5,"x":7,"y":9},{"rot":180,"type":0,"x":6,"y":9},{"rot":180,"type":5,"x":5,"y":9},{"rot":180,"type":-1,"x":4,"y":9},{"rot":180,"type":5,"x":3,"y":9},{"rot":180,"type":2,"x":2,"y":9},{"rot":180,"type":-2,"x":1,"y":9},{"rot":180,"type":-2,"x":0,"y":9},{"rot":180,"type":0,"x":13,"y":10},{"rot":180,"type":5,"x":12,"y":10},{"rot":180,"type":5,"x":11,"y":10},{"rot":180,"type":0,"x":10,"y":10},{"rot":180,"type":2,"x":9,"y":10},{"rot":180,"type":5,"x":8,"y":10},{"rot":180,"type":2,"x":7,"y":10},{"rot":180,"type":5,"x":6,"y":10},{"rot":180,"type":-1,"x":5,"y":10},{"rot":180,"type":5,"x":4,"y":10},{"rot":180,"type":0,"x":3,"y":10},{"rot":180,"type":-2,"x":2,"y":10},{"rot":180,"type":-2,"x":1,"y":10},{"rot":180,"type":-2,"x":0,"y":10},{"rot":180,"type":-2,"x":13,"y":11},{"rot":180,"type":-2,"x":12,"y":11},{"rot":180,"type":5,"x":11,"y":11},{"rot":180,"type":2,"x":10,"y":11},{"rot":180,"type":5,"x":9,"y":11},{"rot":180,"type":2,"x":8,"y":11},{"rot":180,"type":2,"x":7,"y":11},{"rot":180,"type":-1,"x":6,"y":11},{"rot":180,"type":1,"x":5,"y":11},{"rot":180,"type":-1,"x":4,"y":11},{"rot":180,"type":5,"x":3,"y":11},{"rot":180,"type":-2,"x":2,"y":11},{"rot":180,"type":-2,"x":1,"y":11},{"rot":180,"type":-2,"x":0,"y":11},{"rot":180,"type":-2,"x":13,"y":12},{"rot":180,"type":-2,"x":12,"y":12},{"rot":180,"type":-2,"x":11,"y":12},{"rot":180,"type":-2,"x":10,"y":12},{"rot":180,"type":-2,"x":9,"y":12},{"rot":180,"type":-1,"x":8,"y":12},{"rot":180,"type":5,"x":7,"y":12},{"rot":180,"type":5,"x":6,"y":12},{"rot":180,"type":5,"x":5,"y":12},{"rot":180,"type":2,"x":4,"y":12},{"rot":180,"type":2,"x":3,"y":12},{"rot":180,"type":-2,"x":2,"y":12},{"rot":180,"type":-2,"x":1,"y":12},{"rot":180,"type":-2,"x":0,"y":12},{"rot":180,"type":-2,"x":13,"y":13},{"rot":180,"type":-2,"x":12,"y":13},{"rot":180,"type":-2,"x":11,"y":13},{"rot":180,"type":-2,"x":10,"y":13},{"rot":180,"type":-2,"x":9,"y":13},{"rot":180,"type":-2,"x":8,"y":13},{"rot":180,"type":-2,"x":7,"y":13},{"rot":180,"type":-2,"x":6,"y":13},{"rot":180,"type":5,"x":5,"y":13},{"rot":180,"type":2,"x":4,"y":13},{"rot":180,"type":-2,"x":3,"y":13},{"rot":180,"type":-2,"x":2,"y":13},{"rot":180,"type":-2,"x":1,"y":13},{"rot":180,"type":-2,"x":0,"y":13}],"objects":{"IDMax":15,"density":[21,15,24,20,14,5],"dockID":5,"dockRotation":180,"pop_mid":0,"pop_old":0,"pop_young":0,"slider1":0.142857,"toggle1":5,"toggle2":0,"toggle3":0,"toggle4":0}}
