--- examples/Virtual_table/static_examples/cityio_api_v2.json
@@ -1,439 +1,0 @@
-{
-  "error": "",
-  "grid": [
-    [1, 0, 270],
-    [0, 7, 270],
-    [0, 8, 90],
-    [4, 7, 0],
-    [2, 6, 270],
-    [1, 2, 0],
-    [4, 4, 180],
-    [6, 3, 270],
-    [6, 2, 90],
-    [0, 2, 90],
-    [3, 5, 180],
-    [6, 4, 90],
-    [1, 7, 90],
-    [5, 7, 0],
-    [3, 5, 0],
-    [5, 6, 180],
-    [1, 7, 0],
-    [5, 0, 180],
-    [6, 2, 270],
-    [1, 7, 90],
-    [4, 5, 180],
-    [2, 8, 90],
-    [3, 6, 90],
-    [5, 7, 180],
-    [1, 4, 180],
-    [1, 3, 0],
-    [5, 9, 0],
-    [5, 7, 180],
-    [4, 0, 180],
-    [0, 5, 90],
-    [0, 8, 180],
-    [2, 9, 180],
-    [1, 7, 90],
-    [3, 4, 270],
-    [4, 0, 90],
-    [1, 5, 0],
-    [0, 1, 270],
-    [4, 6, 0],
-    [3, 3, 90],
-    [2, 1, 270],
-    [0, 6, 180],
-    [0, 2, 90],
-    [4, 3, 270],
-    [6, 4, 180],
-    [0, 7, 270],
-    [2, 2, 0],
-    [6, 1, 0],
-    [6, 0, 270],
-    [1, 1, 180],
-    [4, 6, 270],
-    [2, 4, 90],
-    [1, 1, 0],
-    [6, 9, 270],
-    [4, 9, 90],
-    [4, 5, 180],
-    [6, 0, 270],
-    [4, 6, 270],
-    [1, 2, 270],
-    [0, 7, 270],
-    [3, 4, 180],
-    [6, 3, 0],
-    [1, 3, 180],
-    [2, 5, 180],
-    [0, 6, 270],
-    [3, 9, 180],
-    [0, 3, 90],
-    [5, 5, 0],
-    [3, 9, 0],
-    [1, 8, 90],
-    [6, 6, 90],
-    [6, 6, 270],
-    [5, 3, 90],
-    [5, 9, 270],
-    [6, 5, 270],
-    [1, 3, 90],
-    [4, 2, 270],
-    [6, 2, 0],
-    [6, 1, 180],
-    [5, 4, 90],
-    [5, 9, 180],
-    [1, 1, 270],
-    [4, 9, 90],
-    [0, 0, 270],
-    [4, 9, 90],
-    [1, 5, 270],
-    [1, 0, 0],
-    [4, 7, 0],
-    [4, 2, 0],
-    [2, 0, 90],
-    [3, 2, 270],
-    [1, 0, 270],
-    [4, 4, 270],
-    [5, 1, 180],
-    [0, 3, 180],
-    [4, 2, 270],
-    [5, 3, 270],
-    [4, 1, 180],
-    [2, 9, 180],
-    [1, 4, 90],
-    [1, 3, 0],
-    [1, 3, 90],
-    [6, 4, 270],
-    [3, 4, 0],
-    [4, 1, 0],
-    [6, 4, 270],
-    [0, 2, 0],
-    [4, 7, 90],
-    [3, 4, 0],
-    [1, 0, 180],
-    [4, 3, 270],
-    [2, 5, 180],
-    [6, 5, 0],
-    [6, 0, 90],
-    [3, 5, 90],
-    [5, 7, 0],
-    [4, 3, 90],
-    [0, 2, 180],
-    [2, 2, 180],
-    [3, 2, 90],
-    [0, 9, 180],
-    [4, 8, 270],
-    [1, 4, 90],
-    [5, 2, 0],
-    [5, 9, 90],
-    [4, 6, 270],
-    [5, 5, 90],
-    [2, 5, 90],
-    [3, 0, 270],
-    [2, 7, 0],
-    [3, 6, 180],
-    [5, 0, 180],
-    [1, 5, 0],
-    [5, 1, 180],
-    [6, 5, 0],
-    [4, 0, 270],
-    [6, 4, 270],
-    [6, 6, 270],
-    [5, 7, 90],
-    [4, 0, 0],
-    [3, 9, 270],
-    [2, 0, 0],
-    [1, 8, 90],
-    [4, 0, 180],
-    [3, 1, 90],
-    [6, 6, 90],
-    [6, 2, 90],
-    [0, 1, 270],
-    [2, 5, 270],
-    [2, 3, 90],
-    [2, 7, 0],
-    [3, 9, 270],
-    [4, 3, 180],
-    [1, 0, 180],
-    [3, 9, 0],
-    [0, 2, 270],
-    [4, 9, 90],
-    [0, 4, 90],
-    [2, 1, 270],
-    [3, 0, 0],
-    [3, 1, 180],
-    [2, 8, 90],
-    [2, 1, 180],
-    [3, 5, 0],
-    [4, 4, 0],
-    [4, 2, 0],
-    [5, 7, 0],
-    [6, 3, 90],
-    [1, 9, 0],
-    [2, 4, 90],
-    [6, 8, 90],
-    [1, 8, 90],
-    [5, 5, 90],
-    [6, 3, 180],
-    [5, 3, 270],
-    [4, 9, 180],
-    [1, 1, 0],
-    [3, 6, 90],
-    [1, 8, 270],
-    [4, 2, 270],
-    [4, 2, 180],
-    [0, 7, 180],
-    [0, 8, 0],
-    [0, 5, 270],
-    [1, 9, 270],
-    [4, 8, 180],
-    [5, 8, 90],
-    [2, 3, 90],
-    [3, 4, 270],
-    [1, 6, 0],
-    [3, 2, 90],
-    [3, 5, 90],
-    [2, 4, 0],
-    [5, 2, 90],
-    [1, 6, 0],
-    [1, 2, 270],
-    [6, 7, 90],
-    [5, 6, 270],
-    [3, 7, 270],
-    [2, 3, 0],
-    [4, 1, 0],
-    [6, 7, 180],
-    [6, 5, 0],
-    [5, 6, 270],
-    [2, 7, 0],
-    [1, 4, 90],
-    [3, 5, 90],
-    [3, 4, 270],
-    [2, 0, 180],
-    [5, 4, 180],
-    [3, 5, 0],
-    [3, 1, 270],
-    [6, 3, 270],
-    [5, 1, 90],
-    [2, 5, 0],
-    [6, 8, 90],
-    [1, 0, 90],
-    [6, 5, 270],
-    [2, 8, 180],
-    [4, 8, 90],
-    [3, 7, 270],
-    [4, 2, 0],
-    [5, 8, 270],
-    [2, 0, 180],
-    [2, 5, 270],
-    [6, 6, 0],
-    [1, 6, 270],
-    [4, 6, 270],
-    [3, 9, 270],
-    [6, 0, 270],
-    [4, 9, 90],
-    [0, 2, 90],
-    [4, 5, 0],
-    [4, 6, 270],
-    [6, 9, 270],
-    [4, 6, 270],
-    [5, 6, 180],
-    [6, 0, 90],
-    [4, 2, 0],
-    [1, 9, 90],
-    [1, 2, 0],
-    [3, 3, 270],
-    [3, 6, 0],
-    [6, 5, 270],
-    [5, 2, 90],
-    [5, 2, 180],
-    [3, 2, 180],
-    [1, 8, 180],
-    [3, 3, 0],
-    [4, 1, 180],
-    [2, 0, 180],
-    [4, 9, 270],
-    [2, 2, 180],
-    [5, 3, 270],
-    [4, 5, 270],
-    [1, 9, 270],
-    [3, 3, 90],
-    [5, 0, 270],
-    [3, 7, 180],
-    [5, 3, 180],
-    [5, 9, 90],
-    [6, 6, 0],
-    [4, 0, 0],
-    [0, 7, 90],
-    [3, 5, 180],
-    [1, 1, 90],
-    [5, 5, 0],
-    [2, 1, 90],
-    [2, 0, 180],
-    [2, 1, 180],
-    [0, 9, 90],
-    [2, 1, 90],
-    [5, 4, 0],
-    [2, 1, 90],
-    [6, 5, 270],
-    [1, 4, 180],
-    [5, 2, 0],
-    [3, 4, 90],
-    [4, 6, 90],
-    [3, 7, 180],
-    [4, 7, 270],
-    [5, 7, 0],
-    [0, 8, 90],
-    [4, 6, 270],
-    [1, 4, 180],
-    [4, 8, 180],
-    [3, 5, 90],
-    [5, 8, 180],
-    [0, 2, 180],
-    [3, 3, 180],
-    [2, 4, 270],
-    [4, 8, 270],
-    [1, 3, 90],
-    [3, 3, 270],
-    [3, 8, 270],
-    [6, 0, 90],
-    [2, 4, 270],
-    [4, 3, 0],
-    [3, 2, 90],
-    [1, 9, 90],
-    [1, 1, 0],
-    [4, 1, 0],
-    [3, 4, 180],
-    [1, 0, 0],
-    [4, 1, 180],
-    [2, 3, 0],
-    [0, 2, 270],
-    [4, 2, 180],
-    [0, 2, 180],
-    [6, 6, 90],
-    [4, 8, 270],
-    [6, 0, 180],
-    [6, 0, 180],
-    [3, 2, 270],
-    [4, 7, 0],
-    [4, 6, 90],
-    [3, 5, 90],
-    [5, 6, 180],
-    [3, 0, 90],
-    [2, 5, 90],
-    [5, 0, 270],
-    [4, 8, 90],
-    [3, 2, 270],
-    [0, 0, 0],
-    [2, 6, 180],
-    [1, 3, 0],
-    [0, 0, 0],
-    [3, 2, 90],
-    [6, 7, 90],
-    [0, 7, 180],
-    [0, 1, 180],
-    [5, 7, 0],
-    [0, 2, 0],
-    [1, 3, 0],
-    [0, 7, 0],
-    [0, 5, 180],
-    [3, 1, 270],
-    [6, 9, 0],
-    [4, 2, 0],
-    [3, 8, 180],
-    [0, 7, 180],
-    [3, 3, 270],
-    [5, 7, 180],
-    [0, 5, 270],
-    [6, 4, 180],
-    [4, 1, 180],
-    [5, 8, 90],
-    [2, 3, 270],
-    [4, 0, 270],
-    [0, 8, 270],
-    [5, 4, 90],
-    [0, 1, 90],
-    [6, 4, 90],
-    [3, 5, 0],
-    [2, 4, 0],
-    [5, 1, 180],
-    [4, 8, 180],
-    [4, 3, 90],
-    [4, 5, 0],
-    [5, 6, 0],
-    [1, 7, 270],
-    [4, 5, 180],
-    [1, 7, 270],
-    [0, 4, 0],
-    [0, 4, 270],
-    [3, 3, 270],
-    [1, 9, 90],
-    [4, 8, 180],
-    [3, 4, 90],
-    [4, 6, 180],
-    [4, 9, 0],
-    [1, 9, 0],
-    [4, 4, 180],
-    [1, 0, 180],
-    [6, 7, 90],
-    [4, 6, 0],
-    [2, 2, 0],
-    [5, 6, 270],
-    [4, 4, 180],
-    [4, 8, 90],
-    [6, 1, 0],
-    [5, 8, 180],
-    [4, 2, 270],
-    [0, 1, 270],
-    [0, 4, 0],
-    [6, 3, 0],
-    [4, 2, 270],
-    [2, 1, 90],
-    [5, 8, 0],
-    [0, 9, 0],
-    [4, 6, 270],
-    [4, 8, 270],
-    [6, 9, 180],
-    [1, 8, 180],
-    [4, 7, 0],
-    [0, 3, 270],
-    [3, 3, 0],
-    [0, 6, 180],
-    [6, 3, 90],
-    [3, 2, 0],
-    [4, 5, 90]
-  ],
-  "id": "",
-  "objects": {},
-  "timestamp": 1528717938368,
-  "meta": { "id": "hnkyCam", "timestamp": 1528717938347, "apiv": 2 },
-  "header": {
-    "spatial": {
-      "physical_longitude": -71.087264,
-      "cellsize": 10,
-      "longitude": -71.087264,
-      "rotation": 0,
-      "nrows": 20,
-      "latitude": 42.360357,
-      "ncols": 20,
-      "physical_latitude": 42.360357
-    },
-    "name": "virtual_table",
-    "block": ["type", "height", "rotation"],
-    "mapping": {
-      "type": {
-        "0": "RL",
-        "1": "RM",
-        "2": "RS",
-        "3": "OL",
-        "4": "OM",
-        "5": "OS",
-        "6": "ROAD"
-      }
-    },
-    "owner": {
-      "name": "Yasushi Sakai",
-      "title": "Research Assistant",
-      "institute": "MIT Media Lab"
-    }
-  }
-}
