--- examples/Virtual_table/vt.json
@@ -1,30 +1,0 @@
-{
-  "table": {
-    "header": {
-      "name": "cityIO_demo_table",
-      "spatial": {
-        "nrows": 25,
-        "ncols": 25,
-        "physical_longitude": 0,
-        "physical_latitude": 0,
-        "longitude": 0,
-        "latitude": 0,
-        "cellSize": 100,
-        "rotation": 0
-      },
-      "owner": {
-        "name": "Ariel Noyman",
-        "institute": "City Science MIT Media Lab"
-      },
-      "block": ["type", "rotation"],
-      "mapping": {
-        "type": ["typeA", "typeB", "typeC", "typeD", "typeE"],
-        "rotation": ["0", "1", "2", "3"]
-      }
-    },
-    "grid": [],
-    "objects": {
-      "cityscopy": {}
-    }
-  }
-}
